// src/theme/tokens.d.ts
export declare const typography: {
    fontFamily: string;
    fontSize: number;
    fontWeightLight: number;
    fontWeightRegular: number;
    fontWeightMedium: number;
    fontWeightSemiBold: number;
    fontWeightBold: number;
    h1: {
        fontFamily: string;
        fontWeight: number;
        fontSize: string;
        lineHeight: number;
        letterSpacing: string;
    };
    h2: {
        fontFamily: string;
        fontWeight: number;
        fontSize: string;
        lineHeight: number;
        letterSpacing: string;
    };
    h3: {
        fontFamily: string;
        fontWeight: number;
        fontSize: string;
        lineHeight: number;
        letterSpacing: string;
    };
    h4: {
        fontFamily: string;
        fontWeight: number;
        fontSize: string;
        lineHeight: number;
        letterSpacing: string;
    };
    h5: {
        fontFamily: string;
        fontWeight: number;
        fontSize: string;
        lineHeight: number;
        letterSpacing: string;
    };
    h6: {
        fontFamily: string;
        fontWeight: number;
        fontSize: string;
        lineHeight: number;
        letterSpacing: string;
    };
    body1: {
        fontFamily: string;
        fontWeight: number;
        fontSize: string;
        lineHeight: number;
        letterSpacing: string;
    };
    body2: {
        fontFamily: string;
        fontWeight: number;
        fontSize: string;
        lineHeight: number;
        letterSpacing: string;
    };
    subtitle1: {
        fontFamily: string;
        fontWeight: number;
        fontSize: string;
        lineHeight: number;
        letterSpacing: string;
    };
    subtitle2: {
        fontFamily: string;
        fontWeight: number;
        fontSize: string;
        lineHeight: number;
        letterSpacing: string;
    };
    button: {
        fontFamily: string;
        fontWeight: number;
        fontSize: string;
        lineHeight: number;
        letterSpacing: string;
        textTransform: string;
    };
    caption: {
        fontFamily: string;
        fontWeight: number;
        fontSize: string;
        lineHeight: number;
        letterSpacing: string;
    };
    overline: {
        fontFamily: string;
        fontWeight: number;
        fontSize: string;
        lineHeight: number;
        letterSpacing: string;
        textTransform: string;
    };
};
/**
 * Shape tokens (border radius)
 */
export declare const shape: {
    borderRadius: number;
    sm: number;
    md: number;
    lg: number;
    xl: number;
    rounded: number;
};
/**
 * Spacing scale
 */
export declare const spacing: number[];